Initial Pi-Capture Commit
Code that should (emphasis on should) be able to run on the ras-pi and record an image every time a specified pin is triggered. This file has not been tested on a Ras-Pi yet so this statement is just a claim so far. The write-testfile.py contined within was used to test the capturing capabilities of the code using a webcam on a laptop.